+++ weblab/accounts/forms.py
@@ -1,5 +1,7 @@
+from braces.forms import UserKwargModelFormMixin
 from django import forms
 from django.contrib.auth import forms as auth_forms
+from django.core.exceptions import ValidationError
 
 from .models import User
 
@@ -27,3 +29,26 @@ class MyAccountForm(forms.ModelForm):
         super().__init__(*args, **kwargs)
         if self.instance.social_auth.exists():
             del self.fields['email']
+
+
+class OwnershipTransferForm(forms.Form):
+    """Used for transferring an existing entity.ownership of an existing entity or dataset """
+
+    email = forms.EmailField(
+        widget=forms.EmailInput(attrs={'placeholder': 'Email address of user'})
+    )
+
+    def _get_user(self, email):
+        try:
+            return User.objects.get(email=email)
+        except User.DoesNotExist:
+            return None
+
+    def clean_email(self):
+        email = self.cleaned_data['email']
+        user = self._get_user(email)
+        if not user:
+            raise ValidationError('User not found')
+
+        self.cleaned_data['user'] = user
+        return email
