+++ weblab/datasets/tests/test_views.py
@@ -281,7 +281,51 @@ class TestTransferDatasetView:
         dataset.refresh_from_db()
         assert dataset.author == other_user
         assert not oldpath.exists()
+        assert not oldpath.parent.exists()
         assert dataset.archive_path.exists()
+
+    def test_datasets_transfer_success_no_file(self, client, logged_in_user, other_user, my_dataset, helpers):
+        dataset = my_dataset
+        assert dataset.author.email == 'test@example.com'
+        response = client.post(
+            '/datasets/%d/transfer' % dataset.pk,
+            data={
+                'email': other_user.email,
+            },
+        )
+        assert response.status_code == 302
+        dataset.refresh_from_db()
+        assert dataset.author == other_user
+
+    def test_datasets_transfer_fail_invalid_user(self, client, logged_in_user, other_user, my_dataset, helpers):
+        dataset = my_dataset
+        assert dataset.author.email == 'test@example.com'
+        response = client.post(
+            '/datasets/%d/transfer' % dataset.pk,
+            data={
+                'email': 'invalid@example.com',
+            },
+        )
+        assert response.status_code == 200
+        dataset.refresh_from_db()
+        assert dataset.author == logged_in_user
+
+    def test_datasets_transfer_fail_same_name(self, client, logged_in_user, other_user, my_dataset, helpers):
+        protocol = recipes.protocol.make()
+        recipes.dataset.make(author=other_user, name='mydataset', visibility='public', protocol=protocol)
+        dataset = my_dataset
+        assert dataset.author.email == 'test@example.com'
+        response = client.post(
+            '/datasets/%d/transfer' % dataset.pk,
+            data={
+                'email': other_user.email,
+            },
+        )
+        assert response.status_code == 200
+        dataset.refresh_from_db()
+        assert dataset.author == logged_in_user
+
+
 
 
 @pytest.mark.django_db
