+++ weblab/entities/forms.py
@@ -182,26 +182,3 @@ class FileUploadForm(forms.ModelForm):
     class Meta:
         model = EntityFile
         fields = ['upload']
-
-
-class EntityTransferForm(UserKwargModelFormMixin, forms.Form):
-    """Used for transferring an existing entity."""
-
-    email = forms.EmailField(
-        widget=forms.EmailInput(attrs={'placeholder': 'Email address of user'})
-    )
-
-    def _get_user(self, email):
-        try:
-            return User.objects.get(email=email)
-        except User.DoesNotExist:
-            return None
-
-    def clean_email(self):
-        email = self.cleaned_data['email']
-        user = self._get_user(email)
-        if not user:
-            raise ValidationError('User not found')
-
-        self.cleaned_data['user'] = user
-        return email
